chore(samples): update samples
