--- custom-login/config-overrides.js
@@ -1,48 +1,0 @@
-/* eslint-disable import/no-extraneous-dependencies */
-const webpack = require('webpack');
-const envModule = require('./env')();
-
-envModule.setEnvironmentVarsFromTestEnv(); // Set environment variables from "testenv" file
-
-process.env.CLIENT_ID = process.env.SPA_CLIENT_ID || process.env.CLIENT_ID;
-process.env.OKTA_TESTING_DISABLEHTTPSCHECK = process.env.OKTA_TESTING_DISABLEHTTPSCHECK || false;
-process.env.USE_INTERACTION_CODE = process.env.USE_INTERACTION_CODE || false;
-
-const env = {};
-
-// List of environment variables made available to the app
-[
-  'ISSUER',
-  'CLIENT_ID',
-  'OKTA_TESTING_DISABLEHTTPSCHECK',
-  'USE_INTERACTION_CODE',
-].forEach((key) => {
-  if (!process.env[key]) {
-    throw new Error(`Environment variable ${key} must be set. See README.md`);
-  }
-  env[key] = JSON.stringify(process.env[key]);
-});
-
-module.exports = {
-  /* eslint-disable no-param-reassign */
-  webpack: (config) => {
-    // Remove the 'ModuleScopePlugin' which keeps us from requiring outside the src/ dir
-    config.resolve.plugins = [];
-
-    // Define global vars from env vars (process.env has already been defined)
-    config.plugins = config.plugins.concat([
-      new webpack.DefinePlugin({
-        'process.env': env,
-      }),
-    ]);
-
-    config.devtool = 'source-map';
-    config.module.rules.push({
-      test: /\.js$/,
-      use: ['source-map-loader'],
-      enforce: 'pre',
-    });
-
-    return config;
-  },
-};
