+++ custom-login/src/Login.jsx
@@ -12,8 +12,9 @@
 
 import React, { useEffect, useRef } from 'react';
 import { useOktaAuth } from '@okta/okta-react';
-import * as OktaSignIn from '@okta/okta-signin-widget';
+import OktaSignIn from '@okta/okta-signin-widget';
 import '@okta/okta-signin-widget/dist/css/okta-sign-in.min.css';
+import logo from './logo.svg';
 
 import config from './config';
 
@@ -36,7 +37,7 @@ const Login = ({ setCorsErrorModalOpen }) => {
       baseUrl: issuer.split('/oauth2')[0],
       clientId,
       redirectUri,
-      logo: `${process.env.PUBLIC_URL}/react.svg`,
+      logo,
       i18n: {
         en: {
           'primaryauth.title': 'Sign in to React & Company',
