+++ okta-hosted-login/package.json
@@ -1,50 +1,28 @@
 {
   "name": "@okta/samples.react.okta-hosted-login",
-  "version": "0.3.0",
   "private": true,
-  "dependencies": {
-    "@okta/okta-auth-js": "^6.0.0",
-    "@okta/okta-react": "^6.4.1",
-    "react": "16.14.0",
-    "react-dom": "16.14.0",
-    "react-router-dom": "5.3.0",
-    "react-scripts": "4.0.3",
-    "semantic-ui-css": "2.4.1",
-    "semantic-ui-react": "2.1.1",
-    "text-encoding": "0.7.0"
-  },
+  "version": "0.3.0",
   "scripts": {
-    "start": "cross-env PORT=8080 react-app-rewired start",
-    "start:browser:none": "BROWSER=none yarn start",
-    "build": "react-app-rewired build",
-    "test": "react-scripts test --watchAll=false",
-    "eject": "react-scripts eject",
-    "lint": "eslint -c .eslintrc.json --ext .js --ext .jsx src/"
+    "prestart": "vite build",
+    "start": "lite-server",
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
+  },
+  "dependencies": {
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-router-dom": "5.2.0",
+    "@okta/okta-auth-js": "^6.2.0",
+    "semantic-ui-css": "2.4.1",
+    "semantic-ui-react": "2.0.3",
+    "text-encoding": "0.7.0",
+    "@okta/okta-react": "^6.4.3"
   },
   "devDependencies": {
-    "h": "1.0.0",
-    "react-app-rewired": "2.1.12",
-    "source-map-loader": "1.1.3",
-    "eslint": "7.32.0",
-    "eslint-config-airbnb": "18.2.1",
-    "eslint-plugin-import": "2.25.4",
-    "eslint-plugin-jsx-a11y": "6.5.1",
-    "eslint-plugin-react": "7.28.0",
-    "eslint-plugin-react-hooks": "4.3.0",
-    "cross-env": "7.0.3",
-    "jest-watch-typeahead": "0.6.5",
-    "dotenv": "8.6.0"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "@vitejs/plugin-react": "^1.0.7",
+    "vite": "^2.8.0",
+    "lite-server": "^2.6.1",
+    "dotenv": "^16.0.0"
   }
 }
