+++ okta-hosted-login/src/Navbar.jsx
@@ -14,6 +14,7 @@ import { useOktaAuth } from '@okta/okta-react';
 import React from 'react';
 import { useHistory, Link } from 'react-router-dom';
 import { Container, Icon, Image, Menu } from 'semantic-ui-react';
+import logo from './logo.svg';
 
 const Navbar = ({ setCorsErrorModalOpen }) => {
   const history = useHistory();
@@ -46,7 +47,7 @@ const Navbar = ({ setCorsErrorModalOpen }) => {
       <Menu fixed="top" inverted>
         <Container>
           <Menu.Item header>
-            <Image size="mini" src={`${process.env.PUBLIC_URL}/react.svg`} />
+            <Image size="mini" src={logo} />
             &nbsp;
             <Link to="/">Okta-React Sample Project</Link>
           </Menu.Item>
